requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
covid==2.4.0
DateTime==4.3
discord==1.0.1
discord-components==0.5.3
discord.py==1.6.0
dnspython==2.1.0
emoji==1.2.0
fast-youtube-search==0.0.8
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gspread==3.7.0
h11==0.12.0
httpcore==0.13.3
httplib2==0.19.0
httpx==0.18.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
mysql-connector-python==8.0.25
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.1
pymongo==3.11.4
pynput==1.7.3
pyparsing==2.4.7
pytz==2021.1
quranpy==1.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
toml==0.10.1
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
youtube-search==2.1.0
youtube-search-python==1.4.5
zope.interface==5.4.0
